api/pixiecore_test.go
package api

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/stretchr/testify/assert"
)

var validMAC = "52:54:00:89:d8:10"

func TestPixiecoreHandler(t *testing.T) {
	store := &fixedStore{
		Machines: map[string]*Machine{validMAC: testMachine},
	}
	h := pixiecoreHandler(store)
	req, _ := http.NewRequest("GET", "/"+validMAC, nil)
	w := httptest.NewRecorder()
	h.ServeHTTP(w, req)
	// assert that:
	// - machine config is rendered as Pixiecore JSON
	expectedJSON := `{"kernel":"/image/kernel","initrd":["/image/initrd_a","/image/initrd_b"],"cmdline":{"a":"b","c":""}}`
	assert.Equal(t, http.StatusOK, w.Code)
	assert.Equal(t, jsonContentType, w.HeaderMap.Get(contentType))
	assert.Equal(t, expectedJSON, w.Body.String())
}

func TestPixiecoreHandler_InvalidMACAddress(t *testing.T) {
	store := &emptyStore{}
	h := pixiecoreHandler(store)
	req, _ := http.NewRequest("GET", "/", nil)
	w := httptest.NewRecorder()
	h.ServeHTTP(w, req)
	assert.Equal(t, http.StatusBadRequest, w.Code)
	assert.Equal(t, "invalid MAC address /\n", w.Body.String())
}

func TestPixiecoreHandler_MissingConfig(t *testing.T) {
	store := &emptyStore{}
	h := pixiecoreHandler(store)
	req, _ := http.NewRequest("GET", "/"+validMAC, nil)
	w := httptest.NewRecorder()
	h.ServeHTTP(w, req)
	assert.Equal(t, http.StatusNotFound, w.Code)
}
